creamos el create de las categorias

+++ lib/src/presentation/pages/admin/category/list/AdminCategoryListItem.dart
@@ -40,10 +40,35 @@ class AdminCategoryListItem extends StatelessWidget {
                 },
                 icon: Icon(Icons.edit)),
             IconButton(
-                onPressed: () {
-                  bloc?.add(DeleteCategory(id: category!.id!));
-                },
-                icon: Icon(Icons.delete)),
+  onPressed: () {
+    showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: Text("Confirmar eliminación"),
+          content: Text("¿Estás seguro de que deseas eliminar esta categoría?"),
+          actions: [
+            TextButton(
+              onPressed: () {
+                Navigator.of(context).pop(); // Cerrar el diálogo sin eliminar
+              },
+              child: Text("Cancelar"),
+            ),
+            TextButton(
+              onPressed: () {
+                bloc?.add(DeleteCategory(id: category!.id!)); // Eliminar categoría
+                Navigator.of(context).pop(); // Cerrar el diálogo después de eliminar
+              },
+              child: Text("Eliminar"),
+            ),
+          ],
+        );
+      },
+    );
+  },
+  icon: Icon(Icons.delete),
+)
+,
           ],
         ),
       ),
